Fixed averageSize int overflow. (#2968)
* Fixed averageSize int overflow.

In the previous variant max average size could be only Integer.MAX_VALUE / 16 = 134217727 ~ 130 Mb String, because of row:
averageSize = (size + 15 * averageSize) / 16;

See the issue: https://github.com/h2database/h2database/issues/2967

+++ h2/src/main/org/h2/mvstore/type/ObjectDataType.java
@@ -1592,7 +1592,7 @@ public class ObjectDataType extends BasicDataType<Object> {
             int size = data.length * 2;
             // adjust the average size
             // using an exponential moving average
-            averageSize = (size + 15 * averageSize) / 16;
+            averageSize = (int) ((size + 15L * averageSize) / 16);
             buff.put((byte) TYPE_SERIALIZED_OBJECT).putVarInt(data.length)
                     .put(data);
         }
@@ -1609,7 +1609,7 @@ public class ObjectDataType extends BasicDataType<Object> {
             int size = data.length * 2;
             // adjust the average size
             // using an exponential moving average
-            averageSize = (size + 15 * averageSize) / 16;
+            averageSize = (int) ((size + 15L * averageSize) / 16);
             buff.get(data);
             return deserialize(data);
         }
